separate vargroup into a module
